Add repositories + JPA mapping + embedded db

+++ src/main/java/com/github/eciuca/todo/services/TodoService.java
@@ -8,7 +8,7 @@ public interface TodoService {
 
     List<Todo> getAllTodos();
 
-    void deleteTodo(int id);
+    void deleteTodo(long id);
 
     Todo createNewTodo();
 }
